refactor(stream): move stream contracts out of adapters
