+++ src/crates/adapters/ai-adapters/Cargo.toml
@@ -11,6 +11,7 @@ crate-type = ["rlib"]
 
 [dependencies]
 anyhow = { workspace = true }
+bitfun-agent-stream = { path = "../../execution/agent-stream" }
 bitfun-core-types = { path = "../../contracts/core-types" }
 chrono = { workspace = true }
 eventsource-stream = { workspace = true }
@@ -22,3 +23,9 @@ serde_json = { workspace = true }
 tokio = { workspace = true }
 tokio-stream = { workspace = true }
 urlencoding = { workspace = true }
+
+[dev-dependencies]
+async-trait = { workspace = true }
+axum = { workspace = true }
+bitfun-events = { path = "../../contracts/events" }
+tokio-util = { workspace = true }
