+++ src/crates/adapters/ai-adapters/tests/common/fixture_loader.rs
