+++ src/crates/adapters/ai-adapters/tests/common/mod.rs
