+++ src/crates/adapters/ai-adapters/tests/common/sse_fixture_server.rs
