+++ src/crates/adapters/ai-adapters/tests/common/stream_test_harness.rs
