+++ src/crates/adapters/ai-adapters/tests/fixtures/stream/anthropic/closed_after_message_delta.sse
