+++ src/crates/adapters/ai-adapters/tests/fixtures/stream/anthropic/empty_thinking_signature_text_and_tool_use.sse
