+++ src/crates/adapters/ai-adapters/tests/fixtures/stream/anthropic/extended_thinking.sse
