+++ src/crates/adapters/ai-adapters/tests/fixtures/stream/anthropic/inline_think_text.sse
