+++ src/crates/adapters/ai-adapters/tests/fixtures/stream/anthropic/interleaved_parallel_tool_use.sse
