+++ src/crates/adapters/ai-adapters/tests/fixtures/stream/anthropic/malformed_content_block_delta.sse
