+++ src/crates/adapters/ai-adapters/tests/fixtures/stream/anthropic/malformed_tool_arguments_extra_brace.sse
