+++ src/crates/adapters/ai-adapters/tests/fixtures/stream/gemini/function_call_string_args.sse
