+++ src/crates/adapters/ai-adapters/tests/fixtures/stream/openai/empty_reasoning_content_text_and_tool_call.sse
