+++ src/crates/adapters/ai-adapters/tests/fixtures/stream/openai/inline_think_text.sse
