+++ src/crates/adapters/ai-adapters/tests/fixtures/stream/openai/interleaved_parallel_tool_args_by_index.sse
