+++ src/crates/adapters/ai-adapters/tests/fixtures/stream/openai/thinking_text_three_tools_with_empty_toolcall_anomaly.sse
