+++ src/crates/adapters/ai-adapters/tests/fixtures/stream/openai/tool_args_split_with_usage.sse
