+++ src/crates/adapters/ai-adapters/tests/fixtures/stream/openai/tool_call_missing_type_field.sse
