+++ src/crates/adapters/ai-adapters/tests/fixtures/stream/openai/tool_call_trailing_empty_args_finish_chunk.sse
