+++ src/crates/adapters/ai-adapters/tests/fixtures/stream/openai/tool_id_only_orphan_filtered.sse
