+++ src/crates/adapters/ai-adapters/tests/fixtures/stream/openai/tool_id_prelude_then_payload_without_id.sse
