+++ src/crates/adapters/ai-adapters/tests/fixtures/stream/openai/two_tools_first_final_chunk_contains_orphan_id_only.sse
