+++ src/crates/adapters/ai-adapters/tests/fixtures/stream/responses/malformed_function_call_arguments.sse
