+++ src/crates/adapters/ai-adapters/tests/stream_processor_anthropic.rs
