+++ src/crates/adapters/ai-adapters/tests/stream_processor_openai.rs
