+++ src/crates/adapters/ai-adapters/tests/stream_processor_tool_arguments.rs
