+++ src/crates/adapters/ai-adapters/tests/stream_replay_regressions.rs
