+++ src/crates/adapters/ai-adapters/tests/stream_test_harness.rs
