+++ src/crates/execution/agent-stream/AGENTS.md
@@ -2,20 +2,22 @@
 
 Scope: this guide applies to `src/crates/execution/agent-stream`.
 
-`bitfun-agent-stream` owns provider stream normalization and replayable stream
-processing contracts. It should preserve provider wire behavior while exposing a
-portable stream surface to higher layers.
+`bitfun-agent-stream` owns provider-neutral stream DTOs, tool-call accumulation,
+and replayable stream processing contracts. Provider wire parsing belongs in
+`src/crates/adapters/ai-adapters`, which converts provider chunks into these
+portable stream contracts.
 
 ## Guardrails
 
 - Do not depend on `bitfun-core`, app crates, Tauri, concrete services,
-  transport adapters, terminal, tool-runtime, or product-domain implementations.
-- Keep provider-specific parsing isolated to stream normalization. Do not add
-  session lifecycle, tool execution, prompt policy, or product orchestration
-  behavior here.
-- Stream fixture changes must preserve ordering, tool-call reconstruction,
+  transport adapters, AI adapters, terminal, tool-runtime, or product-domain
+  implementations.
+- Keep provider-specific SSE or response parsing in `bitfun-ai-adapters`; this
+  crate only owns provider-neutral stream assembly and replay behavior.
+- Do not add session lifecycle, tool execution, prompt policy, or product
+  orchestration behavior here.
+- Stream contract changes must preserve ordering, tool-call reconstruction,
   reasoning/thinking fields, usage accounting, and malformed-chunk handling.
-- New provider quirks need fixture coverage rather than broad catch-all parsing.
 
 ## Verification
 
@@ -23,5 +25,8 @@ portable stream surface to higher layers.
 cargo test -p bitfun-agent-stream
 node scripts/check-core-boundaries.mjs
 ```
+
+When provider fixture parsing changes, run the focused `bitfun-ai-adapters`
+stream tests as well.
 
 For documentation-only changes, run `git diff --check`.
