+++ src/crates/execution/agent-stream/Cargo.toml
@@ -12,7 +12,6 @@ crate-type = ["rlib"]
 [dependencies]
 anyhow = { workspace = true }
 async-trait = { workspace = true }
-bitfun-ai-adapters = { path = "../../adapters/ai-adapters" }
 bitfun-events = { path = "../../contracts/events" }
 futures = { workspace = true }
 log = { workspace = true }
@@ -23,6 +22,4 @@ tokio-util = { workspace = true }
 uuid = { workspace = true }
 
 [dev-dependencies]
-axum = { workspace = true }
-reqwest = { workspace = true }
 tokio-stream = { workspace = true }
